Remove AB3 dependencies in TestUtil
